Added Makefile for assembly and updated README.md
Since Compiling source code, Linking object files and libraries, Testing
and Debugging requires using a whole range of tasks, it was necessary to
choose the right tool to perform these tasks. Of course IDE contains all
of the things you need to develop, compile, link, and debug programs,
but each IDE has a different, and often incompatible, implementation
with other IDEs, and also, due to hiding its implementation, it
interferes with a full understanding of all the processes that take
place. To solve the above problems, a make program was chosen.